Refactor calculate time now returns double

+++ src/main/java/edu/temple/cis/paystation/RateStrategy.java
@@ -5,6 +5,7 @@ interface RateStrategy {
     /**
      * Calculate the time that will be purchased based on a specified rate strategy.
      *
+     * @return
      */
-    int calculateTime(int amount);
+    double calculateTime(int amount);
 }
